backend/requirements.txt
# Backend
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
transformers==4.44.2
torch==2.4.0
torchaudio==2.4.0
TTS>=0.22.0
elevenlabs>=0.2.0
pyttsx3>=2.90
pydub>=0.25.1
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Testing
pytest>=7.0.0
httpx>=0.24.0
